connection: Remove unnecessary '__connman_service_get_index' calls.
In both 'set_default_gateway' and 'unset_default_gateway' there are
unnecessary, redundant, and, per the below, circular calls to
'__connman_service_get_index' to get the network interface index from
the 'service' field of the gateway 'data' parameter passed to these
functions.

However, in 'add_gateway', the gateway data 'index' field is assigned
as follows:

    temp_data->index = index;

where 'index' comes from a parameter passed to 'add_gateway', from
'__connman_connection_gateway_add':

    index = __connman_service_get_index(service);

that was sourced from the network service from which a strong
reference was retained when the gateway data was initialized.

Consequently, the network interface index sought may simply be
accessed from the 'index' field of the gateway data as:

    data->index