connection: Simplify 'set_default_gateway' flow.
Ostensibly over time, the logic of 'set_default_gateway' has become a
bit complicated with early returns, gotos, repeated state checks, and
duplicated function calls.

However, amidst all of that, there is a simpler flow to surface. At
its core, the function is two outer conditional blocks, one for IPv4
and one for IPv6, and three inner conditional blocks within those: one
for a VPN, one for a gateway with an any/unspecified address, and one
for everything else. On success, a call to
'__connman_service_indicate_default' is made; on failure, it is not.

This simplifies the flow of 'set_default_gateway' accordingly,
curtailing the number of early returns, eliminating all of the gotos,
reducing the repeated state checks and function calls to one through
the use of those simplified conditional blocks.