connection: Simplify 'unset_default_gateway' flow.
Ostensibly over time, the logic of 'unset_default_gateway' has become
a bit complicated with early returns and repeated state checks.

However, amidst all of that, there is a simpler flow to surface. At
its core, the function is two outer conditional blocks, one for IPv4
and one for IPv6, and three inner conditional blocks within those: one
for a VPN, one for a gateway with an any/unspecified address, and one
for everything else.

This simplifies the flow of 'unset_default_gateway' accordingly,
curtailing the number of early returns and reducing the repeated state
checks to one through the use of those simplified conditional blocks.