connection: Rename 'find_active_gateway_data'.
This module has a map that maps services to their gateway data and IP
type-specific configuration. That configuration includes a state
variable indicating whether the configuration is actively contributing
to route that has been confirmed (or "activated") by a Routing Netlink
(rtnl) notification from the kernel.

In a multi-technology and -service environment, there are potentially
at least as many gateways as there are technologies and connected
services. Consequently, the function name 'find_active_gateway_data'
is ambiguous in that "which activate gateway data?" is
unspecified. While the purpose of this function remains unclear, the
name is being updated to accurately reflect what it does, which is to
return any active gateway data.