connection: Replace gateway config VPN Boolean.
There is presently overloading of gateway configuration state. That
is, using configuration fields for purposes beyond those for which
they may have originally been intended.

While the current VPN Boolean appears unoverloaded, this migrates it
from a single Boolean to a flag in a 32-bit flags field, allowing for
other future flags that alter how a gateway configuration behaves or
is treated, relative to its state and type.