connection: Replace gateway config active Boolean.
There is presently overloading of gateway configuration state, in
particular, the 'active' field. By "overloading", that is, using the
gateway configuration field for a purpose beyond that for which it may
have been originally intended.

The intent of the 'active' field was to make it a state indicating
that a gateway route, added or deleted, had been confirmed or
"activated" by a Linux Routing Netlink (rtnl) notification via this
module's 'connection_{new,del}gateway' functions.

This returns 'active' to its former and sole role as a representative
of that gateway configuration lifecycle state by making it a state
enumeration, currently with two values, ACTIVE and INACTIVE. In the
future, these will be expanded to further represent all states of the
gateway configuration lifecycle.