connection: Rename 'active_gateway' local.
In '__connman_connection_update_gateway' there is a local variable
inside a loop control block, 'active_gateway'. This renames this local
varable to 'current_gateway' to avoid any confusion with the term
"active" which has implications for the 'ACTIVE' gateway configuration
state, which is definitely not the case for this local.