connection: Refactor '__connman_connection_gateway_add'.
In '__connman_connection_gateway_add', there are only two allowable
values for the 'type' parameter among four total in the enumeration
for 'type'. Error-check against the two invalid values from that
enumeration and return -EINVAL for the invalid values.

In addition, use this check to set 'do_ipv4' and 'do_ipv6' local
Booleans (similar to other functions in the module) that allows for
simplifying the overall logic and flow of the function.