connection: Introduce gateway configuration 'type'.
There is presently overloading of gateway configuration state. That
is, using configuration fields for purposes beyond those for which
they may have originally been intended.

This adds a new gateway configuration field, 'type', an enumeration
indicating the current type or use of the gateway configuration. This
may be used in conjunction with the 'state' and 'flags' fields to
conditionally treat the gateway configuration.