connection: Add support for low-priority default routes.
This adds support for low-priority (that is, metric > 0) default
routes along side the existing high-priority (that is, metric 0)
default route taken on by the default service.

For IPv6, these are handled and managed automatically by the kernel as
part of Router Discovery (RD) Router Advertisements (RAs) and because
link-local addresses and multi-homing are a natural part of IPv6,
nothing needs to be done for these here. Such routes show up in 'ip -6
route show' as:

    default via fe80::f29f:c2ff:fe10:271e dev eth0
        proto ra metric 1024 expires 1622sec hoplimit 64
        pref medium
    default via fe80::f29f:c2ff:fe10:271e dev wlan0
        proto ra metric 1024 expires 1354sec hoplimit 64
        pref medium

For IPv4, largely invented before the advent of link-local addresses
and multi-homing hosts, these need to be fully-managed here and, with
such management, should up in 'ip -4 route show' as low-priority (that
is, metric > 0) default routes:

    default via 192.168.2.1 dev wlan0 metric 4294967295

The other alternative to low-priority routes would be to use "def1"
default routes commonly used by VPNs that have a prefix length of 1
(hence the "def1" name). These would should up as:

    0.0.0.0/1 via 192.168.2.1 dev wlan0
    128.0.0.0/1 via 192.168.2.1 dev wlan0

However, since these require twice the number of routing table entries
and seem no more effective than the low-priority route approach, this
alternative is not used here at present.

The low-priority metric is determined on a per-network interface basis
and is computed by 'compute_low_priority_metric'.

Operationally, down calls from outside this module generally come from
the following three functions:

  1. __connman_connection_gateway_add
  2. __connman_connection_gateway_remove
  3. __connman_connection_update_gateway

and up calls generally come from the following two functions:

  1. connection_newgateway
  2. connection_delgateway

From these five functions above, we are then either attempting to do
the following for a gateway associated with a network service and its
underlying network interface:

  1. Set, or add, the high- or low-priority default route(s).
  2. Unset, or remove, the high- or low-priority default route(s).
  3. Promote the default route from low- to high-priority.
  4. Demote the default route from high- to low-priority.