connection: Introduce gateway config 'ops'
This introduces a gateway configuration 'ops' data structure of IPv4-
and IPv6-specific addressing and routing functions.

With this introduction, all accesses to these functions are done
through a pointer to an "ops" structure. This further neutralizes
IPv4- and IPv6-specific code blocks and clears a path towards their
unification.