[storage] Fix VPN provider not being saved in some cases. Fixes JB#60967
On some cases it seemed that VPN provider settings was not saved when
creating new or the changes to existing one were not stored. Usage of
g_mkdir_with_parents() and better structure to avoid logic errors seemed
to fix the issue.

Cleaned up the code on both provider and service saving and added more
debug and error logging for future use. Changed
__connman_storage_save_provider() to return the result of the operation.