main: Use preprocessor definition rather than 'General' string literal.
This converts a potential run time error due to a mis-spelling of the
"General" configuration group name repeatedly used as a string literal
into a potential compile time error due to a mis-spelling of the
preprocessor definition that now references it.

There is no change in code size to .bss, .data, or .text due to
this change.