[connection] Harmonize setting of VPN gw with regular ones. JB#61260
The process of setting the default gw with other than VPN types
immediately returned after failure and avoided setting the gw then
active. This would create unnecessary noise when interface is put down
and unsetting of gw takes place if the VPNs would still let the active
status to be changed even when setting of gw interface fails.

Also improve the code consistency a bit to have each case to use "goto".