gweb: Support adding/removing routes to DNS servers for query
[gweb] Support adding/removing routes to DNS servers for query. JB#61410

Add support for a callback set for g_web_request_get() to be used for
adding and removing routes to DNS servers when doing the hostname
lookup. This way it can be ensured that the name lookup succeeds as the
routes are set to be using the set inteface since the gweb is using only
the nameservers of that particular service. In some cases the DNS
servers may not be accessible over the currently active network.

The DNS route request is almost identical to the basic route request
happening after the name has been resolved to an IP address. However, in
this case only the index should be needed, as any other address
information is not available at that point. It is intended to use
service lookup based on the index and then utilize service functionality
to setup the routes. Same callback is used for both adding and removal
operations with the set boolean.

The call to dns route callback is made before the hostname lookup and
routes are cleared in the returning function before cancel check in
order to get them always removed even in case the request is cancelled
while waiting for a response.