wispr: Implement DNS route toggling callback function
[wispr] Implement DNS route toggling callback function. JB#61410

Implement callback to be used as DNS route adding function for gweb
prior to making the hostname lookup when making a get request. Simply
get the service based on the index and use service.c functionality to
add the nameservers that are being used by gweb. IP family is handled
internally by service.c determining it from the gateway IP address.