[service] Start online check when upgrading to ready state. JB#61410
Make online check to start always when a service is getting upgraded to
ready state. It is necessary to start this in these cases because
sorting is based also on the state not solely on the preferred list and
relying on the default service getting online check set is not enough.
If there are two services, e.g. mobile data and WLAN connected, first
mobile data that is upgraded/downgraded to ready and then WLAN getting
upgraded to ready state online check is not triggered for the WLAN.