service: Use reference counter for nameserver route add/del
[service] Use reference counter for nameserver route add/del. JB#61410

Implemented a reference counter to be used when adding or removing
nameserver routes for a service. This avoids adding and or removing the
routes twice, or being removed if the routes have already been added.
This can happen in case the service that is about to do online check is
the only active one and already has added nameserver routes then after
name resolving completes the routes would be removed.

Added a forced use in the case when user has added new nameservers via
D-Bus. User set nameservers overrule anything set by other means.