[unit] Add tests for DNS route refcount + fix service test. JB#61410
Add unit tests for DNS route reference counters. This required
adding a hefty amount of stubs for many parts of ConnMan and as a result
fewer sources are needed. The functionality added is a simulation of how
the sequence goes when service goes ready, starts online check and
setups routes for DNS servers while becoming online. In addition the
manual changes done between online check steps is tested by crafting
D-Bus message to do that.