openvpn: Improve compatibility with 2.6 mgmt interface
[openvpn] Improve compatibility with 2.6 mgmt interface. JB#61676

New 2.6 OpenVPN mgmt interface behaves a bit differently depending on
the server that is being connected to. The change has been made on how
the cipher negotiation is done and mgmt interface would be looping the
error up to 128 times in some cases until disconnecting. Therefore, it
is important to react to these errors as connection cannot recover from
it without user interaction.

This change implements checking of these errors (picked from OpenVPN
sources) and reacting to them accordingly. These errors are regarded as
terminal errors so following the approach builtin to ConnMan the
provider must be restarted after each received error.

In addition, authentication error limit (as these errors are prefixed
with AUTH error message) is now being checked within the plugin and if
either the limit is reached or the error is a terminal one ECONNABORT is
passed to the connection callback triggering necessary changes for the
VPN.