vpn-provider: Add auth limit getter and send auth/conn errors to connman
[vpn-provider] Send auth/conn error counts to connman. JB#61635

This change makes the authentication and connection errors, as well as
the authentication error limits passed to connmand. By doing so connmand
can make decisions, such as disable autoconnect when authentication
limit is reached, to avoid the user getting temporary banned by the VPN
provider because or repeteative authentication erros, for instance.

Authentication error limit is sent when creating the connection or when
the value changes. Both errors are sent when they are changed.

Also add getter for authentication limit to be used by plugins.