vpn: Disable autoconnection if ECONNABORTED and auth err limit reached.
[vpn] Disable autoconn if ECONNABORTED and auth err limit reached. JB#61635

ECONNABORTED is from situations where the error is terminal on a VPN and
if the authentication error limit set is reached autoconnection must be
disabled to avoid repeteative attemps which may lead in temporary bans.

This required adding also processing of the authentication limit as well
as authentication and connection error counts.