vpn-provider: Set provider to failed state if auth limit is reached.
[vpn-provider] Set provider to failed state if auth limit is reached. JB#61635

Check with ECONNABORTED that if the authentication limits are reached
set the state to failure instead of disconnecting. This way the error
propageates to user and the error counters can be cleared so it
autoconnection works after user fixes the configuration.

Split the handling of ECANCELED and ECONNABORTED. ECANCELED comes from
canceled agent dialog and any connected VPN should be only disconnected
with that error.