gdhcp: Set pkt_len for packets received in L3 mode
The packet length is not set correctly when an IPv4 L3 ACK is received
from a DHCP server.  This causes the packet to be rejected for having
no MESSAGE_TYPE option and the lease is not renewed at the T1 point
(or any subsequent retry).  Instead, the lease gets re-bound at the T2
point.

Fixes: 58d397ba7487 ("gdhcp: Avoid reading invalid data in dhcp_get_option")