ofono: combine create_device and ready_to_create_device
These two functions are called one after the other, repeating the same
comment in multiple places.  Combine them into a single function that
takes care of checking whether the device can be created, and if so,
creating it.  Move the repeated comment to the new function.

While here, refactor handling of 'ident' which removes the need for the
'out' label.  Also, simplify some logic left over from CDMA support that
was checking whether modem->imsi exists.