gdbus: Ensure proxy method call frees user_data
[gdbus] Ensure proxy method call frees user_data. JB#51243

Make sure that the given user_data is freed always when destroy callback
is defined to avoid any memory leaks. This is because the ownership of
the user_data is passed to the proxy method call, and with these changes
the code conforms to rest of the codebase.

Also call g_dbus_send_message() only when both function and destroy
callback are not given to avoid leaking any memory.