[rtnl] Read bond and dsa device types correctly. JB#51243
This became more visible after 1.34 changes that retries the unknown
uevent type checks and caused the bond device warning to spam the logs.
We have improved the rtnl part to separately read the uevent device type
and dropping the "DEVTYPE=" completely from the line. Upstream does
include the whole read line for checks and, thus, has the +8 for each.
Simply compare the read device type as a whole in the style of our
approach.