[unit] Follow iptables indexes in firewall-iptables test. JB#51243
Use the indexes defined in netfilter_ipv4.h for iptables chains in the
test. This makes the test more consistent with the implementation and
iptables.

Do also some cleanup and optimization for the test code. Make the
internal lists to be NULL'd also by the __connman_iptables_cleanup() to
avoid test crashes.