[wifi] Return -EBUSY for scan request if tethering is on. JB#51243
Upstream WiFi plugin reports -EBUSY when scan is requested while
tethering is enabled. Do the same here to comply with the internal
workings. The actual effect of this is that in both service.c and
technology.c the calls to initiate scan would not ever get to create a
D-Bus reply as the __connman_device_request_hidden_scan() nor
__connman_device_request_scan() would return an ok value. By returning
-EBUSY it will be sent back to the caller as an error.