[firewall] Load iptables extension rules in init. Fixes JB#62248
The iptables extension is the D-Bus API for managing iptables rules via
ConnMan. The rules it has will be loaded at start and before commit
a42c94abcb61063d489ddb7fb5c3453244961b51 they were loaded before
firewall init. This commit returns the functionality to as it was.

The symptom of this was that the policies set in firewall-iptables.c
were overridden by the extension and this is not intended. The reason is
that the common chains are saved by the extension at shutdown, at the
point when ConnMan has cleaned up iptables so all policies would be set
to ACCEPT. And when loaded after firewall the firewall settings would be
ignored.