vpn: Always call the vpn provider connect callback
[vpn] Always call the vpn provider connect callback. JB#51243

In order to track the state and the errors properly add the callback to
every VPN connect call. This was missing for the no daemon type VPNs.

The errors now propagate on this to higher level as well, and does not
make the VPN autoconnection based approaches confused and stuck.