[wifi] Ensure connman_inet_add_to_bridge() gets always called. JB#62287
When tethering is enabled we should always ensure that the WiFi device
is added to the bridge. This was not called because the newlink watch
for the bridge device was missing, resulting the
wifi_device_bridge_check() not being called in these cases, which
eventually gets connman_inet_add_to_bridge() called.

This change adds the newlink watch for the bridge index to monitor
newlink events from kernel netlink. This is done at tether3 stage after
the interface creation is requested from wpa_supplicant via
libgsupplicant as the interface is up at this point.