service: Add check for autoconnect controlled service disconnect
[service] Add check for autoconnect controlled service disconnect. JB#62635

Add a simple function that can be used to define what service types are
controlled by autoconnect off functionality. This can be extended later
to have a config option to control the types.

VPNs are connected on with autoconnect boolean value over D-Bus. Do
disconnect when autoconnect is disabled on a VPN to achieve consistency
in operation. Leave other service types out of the controlled list.