provider: Add support for dual-IP VPNs
[provider] Add support for dual-IP VPNs. JB#62713

Handle both IP family types for a VPN if it, like WireGuard, has them.
Use a boolean array to define AFs for a VPN.

Change the functionality regarding supported families to use a boolean
array defining which address families are supported by a VPN.

Modify the connman_provider_get_family() to be a query type based on the
AF given.